third_party/blink/web_tests/virtual/scoped-custom-element-registry-disabled/external/wpt/custom-elements/registries/Document-customElementRegistry-expected.txt

This is a testharness.js-based test.
[FAIL] customElementRegistry on a document should return window.customElements by default
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] customElementRegistry on a document without a browsing context should return null
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] customElementRegistry on a document of a connected iframe should return contentWindow.customElements
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] customElementRegistry on a document of a disconnected iframe should return contentWindow.customElements
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
Harness: the test ran to completion.
